requirements.txt
streamlit==1.26.0
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.12.3
google-generativeai==0.6.0
